webframeworks/blog_website/requirements_dev.txt
ruff==0.0.285
black==23.7.0
mypy==1.5.1
django-stubs==4.2.3